+++ bin/actions/promote_archive_action.js
@@ -1,15 +1,17 @@
 const { configuration } = require('../core/configuration')
 
+const ReferenceStrategy = require('../strategies/reference_strategy')
 const PackageSystem = require('../support/package_system')
 const FileSystem = require('../support/file_system')
 const _ = require('lodash')
 
 // TODO cache works on one archive at a time, try `all` for every package? or *
 
-module.exports = (specifier, stagingPath) => {
-  const configurationJson = PackageSystem.selectConfiguration()
+module.exports = (specifier) => {
   const paths = configuration.get(`paths`)
-  const archivePath = `${paths.archives}/${specifier.archive}/`
+  const stagingPath = ReferenceStrategy.buildStagingPath(specifier)
+  const configurationJson = PackageSystem.selectConfiguration(stagingPath)
+  const archivePath = ReferenceStrategy.buildArchivePath(specifier, configurationJson)
 
   return FileSystem
     .makeDirectory(`${paths.archives}/`)