+++ bin/packages/folder_package.js
@@ -1,20 +1,13 @@
 const fs = require('fs-extra')
 
-const { configuration } = require('../core/configuration')
-const paths = configuration.get(`paths`)
-
 class FolderPackage {
-  static sendToStaging ({ archive }, cachePath) {
-    const stagingPath = `${paths.staging}/${archive}/`
-
+  static sendToStaging ({ stagingPath }, cachePath) {
     return new Promise((resolve, reject) => {
       fs.copy(cachePath, stagingPath, err => {
         if (err) {
           reject(new Error(err))
         } else {
-          resolve({
-            stagingPath
-          })
+          resolve({})
         }
       })
     })