+++ bin/packages/tar_package.js
@@ -1,14 +1,9 @@
 const FileSystem = require('../support/file_system')
 const tar = require('tar')
 
-const { configuration } = require('../core/configuration')
-const paths = configuration.get(`paths`)
-
 class TarPackage {
-  static sendToStaging (specifier, cachePath) {
+  static sendToStaging ({ stagingPath }, cachePath) {
     return new Promise((resolve, reject) => {
-      const stagingPath = `${paths.staging}/${specifier.archive}/`
-
       return tar.extract({
         cwd: stagingPath,
         file: FileSystem.readPath(cachePath),
@@ -16,9 +11,7 @@ class TarPackage {
         unlink: true,
         strip: 1
       }).then(() => {
-        return {
-          stagingPath
-        }
+        return {}
       })
     })
   }