+++ bin/strategies/reference_strategy.js
@@ -2,15 +2,16 @@ const IOStrategy = require('./io_strategy')
 const path = require('path')
 const _ = require('lodash')
 
-const {
-  configuration
-} = require('../core/configuration')
+const { configuration } = require('../core/configuration')
 
 const Discover = {
   IS_SCOPE: /^@/i,
   IS_VERSION: /^[~^]?\d{1,2}\.\d{1,2}\.\d{1,2}$/i,
-  COMPONENT_NAME: /([a-z0-9-]*).*?\.(?:zip|tar|tgz|gz|tar\.gz)?$/i
+  COMPONENT_NAME: /([a-z0-9-]*).*?\.(?:zip|tar|tgz|gz|tar\.gz)?$/i,
+  FULL_COMPONENT_NAME: /(.*?)\.(?:zip|tar|tgz|gz|tar\.gz)?$/i
 }
+
+// DUPLICATE_SEPERATOR: new RegExp(`${path.sep}+`, `g`)
 
 /**
  * a _reference_ can have a shape of scope, resource, or specifier
@@ -40,75 +41,86 @@ class ReferenceStrategy {
    * @return "reference addendum"
    */
   static normalizeReference (reference, addendum) {
-    return _.trimEnd(`${reference || ''} ${addendum || ''}`)
+    addendum = _.trimStart(addendum || '', path.sep)
+    reference = reference || ''
+
+    return _.trimEnd(`${reference} ${addendum}`)
   }
 
-  static scopeToResource (scope) {
-    const isNotScope = !scope.match(Discover.IS_SCOPE)
+  static scopeToResource (scopeOrResource) {
+    const patternMarkers = configuration.get(`rules.patternMarkers`)
+
+    const isNotScope = _.first(scopeOrResource) !== patternMarkers.source
     if (isNotScope) {
-      return scope
+      return scopeOrResource
     }
 
-    const scopeAspects = IOStrategy.__getScopeAspects(scope)
-    const sources = configuration.get(`sources`)
-    const sourceKey = _.findKey(sources, (_0, sourceKey) => {
-      return scopeAspects.source === sourceKey
-    })
+    const [reference, _ignoreAddendum] = scopeOrResource.split(` `)
+    const [uri, version] = reference.split(patternMarkers.version)
+    const uriAspects = uri.split(patternMarkers.separator)
+    const sourceConversionRule = this.__matchCoversionRule(uri, uriAspects)
 
-    // error
-    if (!sourceKey) {
-      return undefined
+    if (!sourceConversionRule) {
+      return scopeOrResource
     }
 
-    const source = sources[sourceKey]
-    const referenceTemplate = source.reference
+    const { pattern, template, constants } = sourceConversionRule
+    const templateVariables = _.zipObject(pattern.split(patternMarkers.separator), uriAspects)
 
-    // error
-    if (!referenceTemplate) {
-      return undefined
-    }
-
-    // source has username/password
-    // scopeAspects has resource, group, source
-    _.template(referenceTemplate)(_.merge({}, scopeAspects, source))
+    return _.template(template)(_.merge({}, templateVariables, constants, { version }))
   }
 
-  /**
-   *
-   * @param {String} reference a string of the reference-request format
-   * @return {String} a string of the resource-request format
-   */
-  static __getScopeAspects (reference) {
-    const aspects = reference.split(`/`)
-    const hasGroup = _.size(aspects) > 2
+  static __matchCoversionRule(uri, uriAspects) {
+    const patternMarkers = configuration.get(`rules.patternMarkers`)
+    const sources = configuration.get(`sources`)
+    const scope = uriAspects[0] = (_.first(uriAspects) || '').substring(1)
 
-    return {
-      resource: _.nth(aspects, hasGroup ? 1 : 2),
-      group: hasGroup ? _.nth(aspects, 1) : ``,
-      source: _.nth(aspects, 0)
-    }
+    return _.find(sources, ({ pattern, reference }, sourceKey) => {
+      return scope === sourceKey
+    })
   }
 
   static resourceToSpecifier (resource) {
+
+    // console.log(configuration.get())
+
+    const versionMarker = configuration.get(`rules.patternMarkers.version`)
+    const stagingFolder = configuration.get(`paths.staging`)
+
     const [reference, addendum] = resource.split(` `)
-    const [uri, version] = reference.split(`#`)
-    const archive = this.__findComponent(uri, addendum)
+    const [uri, version] = reference.split(versionMarker)
+    const fullURI = `${reference || ''}/${addendum || ''}`
+    const fullFolderURI = this.__findPathAspect(fullURI, Discover.FULL_COMPONENT_NAME)
+    const archive = this.__findPathAspect(fullURI, Discover.COMPONENT_NAME)
+
+    const stagingPath = `${stagingFolder}/${fullFolderURI}/`
 
     return {
       version: version || `master`,
-      archive,
+      stagingPath,
       addendum,
+      archive,
       uri
     }
   }
 
-  static __findComponent (reference, addendum) {
-    const fullURI = `${reference || ''}/${addendum || ''}`
+  static __findPathAspect (fullURI, regexFind) {
     const lastAspect = _.last(_.compact(fullURI.split(path.sep || `/`)))
-    const extractZipName = Discover.COMPONENT_NAME.exec(lastAspect)
-    const archiveName = (extractZipName) ? _.last(extractZipName) : lastAspect
+    const extractZipName = regexFind.exec(lastAspect)
+    const pathAspect = (extractZipName) ? _.last(extractZipName) : lastAspect
 
-    return archiveName
+    return pathAspect
+  }
+
+  static buildArchivePath (specifier, archiveManifest) {
+    const paths = configuration.get(`paths`)
+    const archiveName = (archiveManifest.name) ? archiveManifest.name : specifier.name
+
+    return `${paths.archives}/${archiveName}/`
+  }
+
+  static buildStagingPath (specifier) {
+    return specifier.stagingPath
   }
 }